bad GCC, down boy.

+++ src/Interface/TextList.cpp
@@ -1094,7 +1094,7 @@ void TextList::scrollTo(size_t scroll)
 {
 	if (!_scrolling)
 		return;
-	_scroll = std::max(0u, std::min(_rows.size() - _visibleRows, scroll));
+	_scroll = std::max((size_t)(0), std::min(_rows.size() - _visibleRows, scroll));
 	draw(); // can't just set _redraw here because reasons
 	updateArrows();
 }
